Test. This one should fail.

+++ gpytorch/kernels/half_matern_kernel.py
@@ -8,4 +8,4 @@ from .. import settings
 class HalfMaternKernel(MaternKernel):
     def forward(self, x1, x2, **params):
         base_out = super().forward(x1, x2, **params)
-        return HalfNonLazyTensor(base_out)
+        return HalfNonLazyTensor(base_out.half().float())
